deprecate arguments of execute function (#5984)
**Context:**

deprecate the following arguments from the`qml.execute` function as part
of the 0.38 deprecations cycle.

- `expand_fn`: Specifying it allows the user to not do the preprocessing
and validation the device said needed to be done. This is dangerous, and
could lead to incorrect results or unmanaged errors.
- `max_expansion`: It allows users to not reach the required target
gateset of the device, potentially leading to incorrect results or
unmanaged errors.
- `override_shots`: Setting the shots on the tape(s) to be executed
provides greater control and flexibility.
- `device_batch_transform`: Setting `device_batch_transform=False` opens
up the user to submitting something unvalidated and unsupported to the
device.

**Description of the Change:**

Several argument in ``qml.execute`` have been deprecated. 

[sc-67226]
[sc-67230]
[sc-67225]
[sc-67224]