Update orbital order in qubit observable (#5950)
**Context:**
The `fermionic_hamiltonian` function returns fermionic terms in a sorted
order following the order of spin orbitals. The construction of the
qubit operator in `qubit_observable` follows the order of fermionic
terms and in very occasional edge cases gives a Hamiltonians with a
non-trivial wire order, which might be misleading for some users. This
PR sorts the order of the terms in the qubit operator according to the
orbital order.

**Description of the Change:**
The `qubit_observable` function is updated to sort qubit terms according
to the orbital order.

**Benefits:**
Molecular Hamiltonians will always have a "trivial" wire order.  

[sc-67935]