Add exponential fitting for ZNE (#5972)
**Context:** As part of increasing the utility of ZNE within Catalyst,
we'd like to add the option to use exponential fitting with ZNE. The
issue

**Description of the Change:** A function which is written using JAX is
added which performs an exponential fit of data, and evaluates said fit
at 0.

**Benefits:** Increased capabilities for users to play with. This is
especially important given the benefits of tuning ZNE to the problem and
backend at hand.

**Possible Drawbacks:** The current implementation transforms data using
the natural logarithm $\ln$ to linearize the data so that a linear fit
can be applied to learn the necessary parameters (see section below on
the exponential model). Once they transformed parameters are learned,
they can be converted back into those from the exponential model quite
simply. The major drawback of this approach is that they expectation
values have to be shifted to be in the domain of $\ln$ (i.e. $> 0$), or
thrown away if small enough (see $\varepsilon$ `eps` parameter
introduced in `expontial_extrapolate`).

**Related GitHub Issues:**
https://github.com/PennyLaneAI/catalyst/issues/754

**Exponential model:** Assuming an exponential model with asymptote
$y(x) = A\mathrm{e}^{Bx} + C$ we must first turn this into a linear
equation to be able to apply a linear fit.

$$y(x) = A\mathrm{e}^{Bx} + C$$

$$y(x) - C = A\mathrm{e}^{Bx}$$

$$\ln(y(x) - C) = \ln(A) + Bx$$

$$\tilde{y}(x) = \tilde{A} + Bx$$

Once $\tilde{A}$ is learned, the hard part is over as the ZNE value is
"simply" $y(0) = A + C$.

---------

Co-authored-by: Cristian Emiliano Godinez Ramirez <57567043+EmilianoG-byte@users.noreply.github.com>