Allow `qml.for_loop` and `qml.while_loop` to fallback to the Python interpreter if a compiler is not available. (#6014)
**Context:**

In Catalyst, `catalyst.for_loop` and `catalyst.while_loop` will
seamlessly fallback to Python interpretation when `@qjit` is not
present. However, `qml.for_loop` and `qml.while_loop` will instead raise
an error when `@qjit` is not present, which is not ideal.

**Description of the Change:**

When `@qjit` is not present, `qml.for_loop` and `qml.while_loop` will
now simply execute the for/while loop in standard Python.

**Benefits:**

- This will allow the same code to work with and without `@qjit`,
without needing to modify the workflow.

- This will also allow templates and operation decompositions to use
`qml.for_loop` and `qml.while_loop` internally, with no changes for
standard Python execution, but with loops captured automatically when
using `@qjit`.

**Possible Drawbacks:**

- Not a drawback per se, more of a note, but the `ForLoopCallable` and
`WhileLoopCallable` classes introduced in this PR might be needed (and
extended) to support capturing for loops/if statements as JAXPR
primitives(?). They are based on the equivalent classes from Catalyst.

**Related GitHub Issues:** n/a

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>