Move `qml.device` to `qml.devices` (#6030)
**Context:**
The root `__init__` file is bloated with device instantiation code
through the `qml.device` function and its helper functions. It is also
contaminating the namespace with unrelated third-part packages imports
needed for its functionality. The code doesn't need to live in this file
since we have a dedicated `qml.devices` module now.

**Description of the Change:**
The relevant code is moved to the `__init__` of the `qml.devices` module
and reference in the root `__init__` file and the docs are adapted. The
imports are also pushed inside the relevant functions to avoid
contaminating the namespace.

**Benefits:**
More localized functional scopes.

**Possible Drawbacks:**
None

[[sc-65678](https://app.shortcut.com/xanaduai/story/65678)]