Updates to lie_closure (#6023)
Fix https://github.com/PennyLaneAI/pennylane/issues/6018 and do some
other maintenance upgrades

Aim is to allow lie closure to reliably handle cases where sums of
paulis are input (e.g. Heisenberg model)

The problem was that sometimes a commutator would be 0 but require a
simplify for that to show. Fixed that.