Deprecate the internal switching of a backpropagation device (#6032)
**Context:**

The internal switching of one device for another
backpropagation-compatible device was developed to allow the
substitution of `default.qubit.legacy` with either
`default.qubit.autograd`, `default.qubit.jax`, `default.qubit.torch`, or
`default.qubit.tf`. This logic has caused a lot of pain over the years,
and all five of those devices are now deprecated. In order to complete
the removal of the 5 legacy devices, we need to also deprecate and
remove the backprop device switching logic.

Otherwise, it would be very difficult and painful to test the switching
of a backpropagation device.

**Description of the Change:**

Deprecates the switching of a device with a different device. A warning
is only raised when the device is not `default.qubit.legacy`, as
default.qubit.legacy raises it's own deprecation warnings.

**Benefits:**

We can finally clean up QNode 🎉 💥 

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: Shiro-Raven <exclass9.24@gmail.com>
Co-authored-by: Thomas R. Bromley <49409390+trbromley@users.noreply.github.com>
Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>