Registers feature (#5957)
**Context:**
Although pennylane allows the definition of wires, it doesn't easily
support the act of creating whole registers (e.g. some ancillary
register and measurement register).

**Description of the Change:**
To make this process easier, this PR implements a new function called
`qml.registers`, which takes in a dictionary as input, whose values are
integers (number of qubits in the register). `qml.registers` will return
a dictionary of Wires instances, where the key is the register name and
the value is the Wires instance. `qml.registers` will also take, as
input, nested registers of registers.

**Benefits:**
Implement registers much easier/faster on an abstract level. No need to
write a for loop just to set up your Wires.

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: Thomas R. Bromley <49409390+trbromley@users.noreply.github.com>
Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>
Co-authored-by: soranjh <40344468+soranjh@users.noreply.github.com>
Co-authored-by: Utkarsh <utkarshazad98@gmail.com>