Capture the adjoint transform into jaxpr (#5966)
**Context:**

We have an ongoing experimental project to enable capture of quantum
functions into a jaxpr representation.

While we can now capture any operator or measurement process, we want to
also be able to natively capture certain kinds of qfunc transforms; the
first of which is `adjoint`.

**Description of the Change:**

Adds information to `capture/explanations.md` about nested jaxpr.

Adds a capture-compatible implementation of the adjoint qfunc transform.

**Benefits:**

**Possible Drawbacks:**

I've gone back and forth about where to place the definition of the
adjoint primitive. Placing it in `capture` keeps it better isolated from
normal pennylane development. Most developers wouldn't see the capture
implementation details. But placing it in `capture` would also cause
circular dependency issues.

We are getting closer to a place where more PL developers will need to
be familiar with the basics of program capture, and so shouldn't be
scared off by jax primitives. So placing it in the normal source code is
good for getting everyone used to it.

Originally, I implemented this task with a `bind_nested_plxpr`
transform-transform that abstracted away a lot of the details about
creating primitives. But I now think that was premature optimization.
While many details are similar, the different transforms have enough
slight differences that we should have custom implementations for each
right now. We can abstract away patterns later.

**Related GitHub Issues:**

[sc-68084]

---------

Co-authored-by: Thomas R. Bromley <49409390+trbromley@users.noreply.github.com>
Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>
Co-authored-by: David Ittah <dime10@users.noreply.github.com>
Co-authored-by: David Wierichs <david.wierichs@xanadu.ai>