Removing `semantic_version` from the list of required packages (#5836)
**Context:** After #5769 and #5754, we can potentially replace all the
occurrences of the `semantic_version` package with `packaging` in the
PennyLane source code, and remove `semantic_version` from the required
Python packages. A similar attempt was performed in #5782, but it was
reverted due to some errors in the plugin-matrix tests.

**Description of the Change:** We replace the remaining occurrences of
`semantic_version` with `packaging` and remove `semantic_version` from
the required packages in PennyLane. With respect to #5782, now the
plugins can be correctly instantiated as:

```
dev = qml.device("qiskit.aer", wires=2)

dev = qml.device("cirq.simulator", wires=1)
```
etc.

Specifically, the reason was the usage of `packaging` to replace
`semantic_version`in the `pennylane/__init__.py` file. There was an
incompatibility between the workflow of the two packages. Right now, the
replacement has been performed to maintain the workflow. For example,

```
from packaging.specifiers import SpecifierSet, Specifier
from packaging.version import Version

pennylane_requires = ">=0.37.0"
version = "0.38.0.dev0"
>>> Version(version) in SpecifierSet(pennylane_requires)
False
>>> Version(version) in SpecifierSet(pennylane_requires, prereleases=True)
True
```

**Benefits:** One package less to keep track of in the dependencies.

**Possible Drawbacks:** Errors in the plugin-text matrix. If they
happen, we'll react as quickly as possible.

**Related GitHub Issues:** None.

**Related Shortcut Stories** [sc-64659]