Incorporate the 'captured' and 'non-captured' versions of `qml.cond` into `CondCallable` (#6063)
**Context:** For 'historical reasons', the captured version (see
sc-66774) and the non-captured version (see sc-69157) of `qml.cond` have
a different style, which is rather uneven.

The purpose of this story is to incorporate them in such a way that the
implementation is more similar to the one in Catalyst. Or, to make
another example, the implementation of `qml.for_loop` in PL (see
sc-66736 and sc-69432). The purpose is not to change/add
functionalities, but rather to unify the code structure so that it is
more elegant.

**Description of the Change:** As above.

**Benefits:** Better and cleaner structure, more similar to Catalyst.

**Possible Drawbacks:** None that I can think of right now directly due
to this PR (we just moved some code).

**Related GitHub Issues:** None

**Related Shortcut Stories:** [sc-70342]