Allow `qml.cond` to fallback to the Python interpreter if a compiler is not available and there are no MCMs (#6016)
**Context:**

In Catalyst, `catalyst.cond` will seamlessly fallback to Python
interpretation when `@qjit` is not present. However, `qml.cond` will
instead raise an error when `@qjit` is not present and the condition is
not a mid-circuit measurement, which is not ideal.

**Description of the Change:**

- When `@qjit` is not present and the conditional does not contain a
mid-circuit measurement, `qml.cond` will now simply execute the
conditional in standard Python.

- `qml.cond` can be optionally used in decorator mode (matching the
behaviour of `catalyst.cond`).

**Benefits:**

- This will allow the same code to work with and without `@qjit`,
without needing to modify the workflow.

- This will also allow templates and operation decompositions to use
`qml.cond` internally, with no changes for standard Python execution,
but with conditionals captured automatically when using `@qjit`.

**Possible Drawbacks:**

- Not a drawback per se, more of a note, but the `CondCallable` class
introduced in this PR might be needed (and extended) to support
capturing for conditional statements as JAXPR primitives(?). It is based
on the equivalent classes from Catalyst.

- Introduces an additional decorator UI. This is not documented for now,
but it will easily allow existing Catalyst code to port from using
`catalyst.cond` to `qml.cond`.

**Related GitHub Issues:** n/a

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>