Add support for mid-circuit measurement program capture (#6015)
**Context:**
This PR is following up on the new program capture effort by adding
support for capturing mid-circuit measurements. Currently, mid-circuit
measurements can be captured, classically processed, and executed, with
specific constraints, but cannot be used with `qml.cond`

**Description of the Change:**
* Add primitive for capturing `qml.measure`. This primitive's concrete
implementation is basically the same as the original, where it queues a
`MidMeasureMP` and returns a `MeasurementValue`, but its abstract eval
returns a ~boolean~ integer tracer. Follow up work will take advantage
of this.

**Benefits:**
* Mid-circuit measurements can be captured
* Arbitrary classical processing of MCMs can be captured. This includes
functionality that is already available, as well as new functionality
such as operations like `jnp.sin(m)`, `jnp.array([m1, m2])`, and even
`jnp.max(jnp.array([m1, m2])`! All of this can be captured, but not
executed.
* Using mid-circuit measurements as gate parameters, eg. `qml.RX(m, 0)`,
can be captured, but not executed.
* Circuits containing mid-circuit measurements can be executed with
`qml.capture.enable()`. This only works when the circuit does not
classically process MCMs or applies `qml.cond` with MCMs in the
predicate.
* Circuits collecting statistics on mid-circuit measurements or
sequences of mid-circuit measurements can be captured and executed.