[Closes #6047] Implements custom `compute_decomposition` method in `Hermitian` (#6062)
**Context:**

Prior to this fix, the `compute_decomposition` method of `Hermitian`
returned a `qml.DecompositionUndefinedError` (due to its `Operator` base
class inheritance). The desired fix, as laid out by the issue, was to
use the existing `pauli_decompose` method to perform the desired
decomposition.

**Description of the Change:**

Implemented a custom `compute_decomposition` static method in
`Hermitian` to have it call the existing `pauli_decompose` method.

<ins>Examples</ins>

```python
>>> op = qml.X(0) + qml.Y(1) + 2 * qml.X(0) @ qml.Z(3)
>>> op_matrix = qml.matrix(op)
>>> qml.Hermitian.compute_decomposition(op_matrix, wires=['a', 'b', 'aux'])
[(
      1.0 * (I('a') @ Y('b') @ I('aux'))
    + 1.0 * (X('a') @ I('b') @ I('aux'))
    + 2.0 * (X('a') @ I('b') @ Z('aux'))
)]

>>> op = np.array([[1, 1], [1, -1]]) / np.sqrt(2)
>>> qml.Hermitian.compute_decomposition(op, wires=0)
[(
      0.7071067811865475 * X(0)
    + 0.7071067811865475 * Z(0)
)]
```

 ⏰ *Benchmarking Performance* ⏰

Performance was benchmarked as a function of matrix size (see figure
below). I heuristically chose to raise an inefficiency warning if the
decomposition time would take longer than a second (open to suggestions
🙂). Any observable with more than seven wires will raise a user warning
indicating that the decomposition may be inefficient.

I've also plotted the theoretical time complexity of the
`pauli_decompose` method and the data is in agreement.


![benchmark.pdf](https://github.com/user-attachments/files/16459301/benchmark.pdf)

**Benefits:**

Can now compute the decomposition of any Hermitian matrix.

**Possible Drawbacks:**

N/A

**Related GitHub Issues:**

Closes #6047

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>
Co-authored-by: Alex Preciado <alex.preciado@xanadu.ai>