Update `Hermitian.decomposition` test to use `qml.matrix` (#6095)
After the merge of #6062 , the legacy op math tests in the plugin test
matrix fail (see
[here](https://github.com/PennyLaneAI/pennylane/actions/runs/10362479768/job/28685492331)),
since the test uses the `Operator.matrix` method to verify the
decomposition, but the decomposed operator is a legacy `Hamiltonian`
with new op math disabled, which doesn't have a matrix. This PR just
updates the test to use `qml.matrix` instead of `Operator.matrix`.