Add diagonalize_measurements transform (#5829)
**Context:**
This transform diagonalizes measurements on a tape. By default, it
diagonalizes all measurements, but generally this can be done more
efficiently using the `pauli_rep` of the observables (next PR). This
transform is primarily useful for the added functionality of only
diagonalizing _some subset_ of {X, Y, Z and Hadamard}, for backends that
support only some of these. This is based on requirements established
for Catalyst backends.

**Description of the Change:**
The `diagonalize_tape_measurements` transform is added. If the backend
supports some non-measurement basis observables, we can diagonalize only
a subset by passing `supported_base_obs`:

```
>>> measurements = [qml.expval(qml.X(0) @ qml.Y(1)), qml.var(qml.X(0))]
>>> tape = qml.tape.QuantumScript([], measurements=measurements)
>>> tapes, _ = diagonalize_tape_measurements(tape, supported_base_obs =["PauliY", "PauliZ"])
>>> tapes[0].measurements
[expval(Z(0) @ Y(1)), var(Z(0))]
```

The transform tracks which observables and wires are already
encountered, and avoids applying diagonalizing gates more than once for
the same wire:

```
>>> tapes[0].operations
[Hadamard(wires=[0])]
```

**Benefits:**
We can diagonalize a subset of observables on a tape.

**Possible Drawbacks:**
- If we are generally diagonalizing (the most common use for
diagonalization), for larger observables performance doesn't scale well
compared to a pauli_rep based approach. We will also implement this
(followup PR).

- This converts all observables to Z instead of to eigvals. It would be
nice to have the option to chose either one, but that is out of scope
and would need to be a future improvement.

---------

Co-authored-by: Astral Cai <astral.cai@xanadu.ai>
Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Thomas R. Bromley <49409390+trbromley@users.noreply.github.com>