Tree-traversal iterative design & analytic-mode (#5868)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      test directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `doc/releases/changelog-dev.md` file,
summarizing the
      change, and including a link back to the PR.

- [x] The PennyLane source code conforms to
      [PEP8 standards](https://www.python.org/dev/peps/pep-0008/).
We check all of our code against [Pylint](https://www.pylint.org/).
      To lint modified files, simply `pip install pylint`, and then
      run `pylint pennylane/path/to/file.py`.

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
The tree-traversal algorithm was first implemented using a recursive
approach for simplicity, but this potentially requires very deep stack
calls for circuits with a lot of MCMs. Tree-traversal could also support
analytic execution, optionally with a probability cutoff to ignore
branches.

**Description of the Change:**
Implement tree-traversal using an iterative approach.
Support `shots=None`.
Some profiling and optimization.

**Benefits:**
Remove system call increasing the default recursion call limit.

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-65241]

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>
Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Matthew Silverman <matthews@xanadu.ai>