Change strategy for determining RC branch version (#6168)
We originally used to end the dev release version with `dev0`. However,
since the nightly releases, the dev version number increases everyday.
This change broke the RC sync to master action, which would check for
the RC branch's existence using the version number from `master`. I
updating the step that computes the RC branch's name to use a regex
instead.