Update stable dependencies on schedule (#6164)
As name says. Currently, every time stable dependencies change (every
time a PR is merged to master), an automatic PR is opened to update the
dependencies. It can be quite annoying to merge a PR and then another
one gets opened immediately afterwards. Thus, this changes the test
workflows to run them on a weekly schedule, and a PR to update stable
dependencies is only opened if the test workflow is triggered by the
schedule.

All changes have been implemented by @timmysilv . I'm just the
messenger.

---------

Co-authored-by: Matthew Silverman <matthews@xanadu.ai>