Minor fixes for 0.38 (#6163)
Minor changes for QA for 0.38:

* Trailing whitespace removed from changelog
* Change `qml.cond`'s docs to stop using `np.array(<value>,
requires_grad=True)` as `np` should be vanilla numpy, not autograd.
Additionally, the use of `pennylane.numpy` does not add anything
meaningful to the examples in the docs.
* I also also added a note to the `qml.cond` docs about falling back to
python control flow when `qml.qjit` is not active.
* Fix whitespace in `LinearCombination` docs.
* Fix formatting and type hints in docs for functions related to
`group_observables`.
* Add license note to `diagonalize_measurements` and wrap some stuff in
inline code blocks in the doc string.
* Minor doc fix to `single_qubit_fusion`.
* Minor formatting and code correction to `split_to_single_terms`'s
docstring
* Minor doc fix for `qml.UCCSD`.
* Squeeze batch dimension of inputs to a function given as input to
`qml.fourier.coefficients`. The squeezing was being done by
`broadcast_expand` before, which is no longer the case.