fix renormalization issue (#6190)
Normalization errors were occuring when the state was jax and x64 mode
was not enabled. Simply renormalizing the state did not help, as the
norm only differed from 1 by machine precision.


This PR simply defers sampling to jax whenever the state is jax. If a
prng key is not provided, we just pull a seed from the numpy rng.
Pulling a prng seed from the numpy rng will not be particularily
jit-able, but if the user wants it to be jittable, they can provide a
prng key instead.

Fixes #6100 [sc-71361]

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>