Miscellaneous updates to CI workflows (#6214)
* Add argument to dependency installation workflow to install Catalyst
from nightly builds and set it to `true` for external tests (tests that
use catalyst)
* Add argument to test workflow for storing durations file as an
artifact and set it to `true` for core, jax and tf tests. The argument
is for defining a file path to an already existing durations file. If
provided, we use the durations file for splitting tests among the
runners, and then also store durations. Thus, we will always store
durations and create artifacts for core, tf, and jax tests. I'm not sure
if this impacts the pytest runtime significantly or not. It can be
changed if needed.
* I set the argument to `false` by default so that we can selectively
set it to true when jobs get very unbalanced again.
* Changed `max_parallel` for device tests to 2, so up to 2 device tests
can run in parallel.

[sc-72850]