[Program Capture] Capture & execute `qml.grad` in plxpr (#6120)
**Context:**
The new `qml.capture` module does not support differentiation yet.

**Description of the Change:**
This PR takes the first step towards differentiability in plxpr.
It adds the capability of capturing `qml.grad` as a "nested jaxpr"
primitive.
When executing the captured program, `qml.grad` is essentially changed
to `jax.grad`, because executing Autograd autodifferentiation within the
Jaxpr ecosystem is not sensible.

**Benefits:**
Capture first differentiation instructions

**Possible Drawbacks:**
The current implementation requires a `jvp` construction for every
evaluation of a QNode gradient. This means that this JVP function is
reconstructed for every evaluation call, if I'm not mistaken, making the
code significantly less performant with `capture` than without. Of
course, the longer term plan is to process the plxpr into lower-level
code by lowering the `grad` primitive itself, in which case this problem
goes away.
A similar redundancy is implemented in `QNode`: Whenever a `qnode`
primitive is evaluated, a new `QNode` is created (and only ever
evaluated once). This disables caching, for example, unless a cache is
passed around explicitly.

**Related GitHub Issues:**

[sc-71858]

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>