PennyLane is compatible with JAX 0.4.28 (#6255)
**Context:** As part of the effort to make PL compatible with Numpy 2.0
(see #6061), we need to upgrade JAX to 0.4.26+ since such a version
introduced the support for Numpy 2.0. We opted for JAX 0.4.28 since it
is the same version used by Catalyst.

**Description of the Change:** As above.

**Benefits:** PL is compatible with Numpy 2.0 and Jax 0.4.28.

**Possible Drawbacks:** 

- From JAX 0.4.27, in `jax.jit`, passing invalid static_argnums or
static_argnames now leads to an error rather than a warning. In PL, this
breaks every test where we set `shots` in the `QNode` call with
`static_argnames=["shots"]`. At this stage, we decided to mark such
tests with `pytest.xfail` to allow the upgrade.

**Related GitHub Issues:** None.

**Related Shortcut Stories**: [sc-61389]

---------

Co-authored-by: dwierichs <david.wierichs@xanadu.ai>